+++ src/Nito.Comparers.Core/ComparerBuilderFor.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using Nito.Comparers.Util;
 
+#pragma warning disable IDE0060
+
 namespace Nito.Comparers
 {
     /// <summary>
@@ -36,6 +38,13 @@ namespace Nito.Comparers
         public static IFullComparer<T> Default<T>(this ComparerBuilderFor<T> @this) => (IFullComparer<T>)ComparerHelpers.NormalizeDefault<T>(null);
 
         /// <summary>
+        /// Gets a natural string comparer, which treats numeric sequences (0-9) as numeric.
+        /// </summary>
+        /// <param name="this">The comparer builder.</param>
+        /// <param name="comparison">The comparison type used to compare the text segments of the string (not used for numeric segments).</param>
+        public static IFullComparer<string> Natural(this ComparerBuilderFor<string> @this, StringComparison comparison) => new NaturalStringComparer(comparison);
+
+        /// <summary>
         /// Creates a key comparer.
         /// </summary>
         /// <typeparam name="T">The type of object being compared.</typeparam>
@@ -46,7 +55,7 @@ namespace Nito.Comparers
         /// <param name="specialNullHandling">A value indicating whether <c>null</c> values are passed to <paramref name="selector"/>. If <c>false</c>, then <c>null</c> values are considered less than any non-<c>null</c> values and are not passed to <paramref name="selector"/>. This value is ignored if <typeparamref name="T"/> is a non-nullable type.</param>
         /// <param name="descending">A value indicating whether the sorting is done in descending order. If <c>false</c> (the default), then the sort is in ascending order.</param>
         /// <returns>A key comparer.</returns>
-        public static IFullComparer<T> OrderBy<T, TKey>(this ComparerBuilderFor<T> @this, Func<T, TKey> selector, Func<ComparerBuilderFor<TKey>, IComparer<TKey>> comparerFactory, bool specialNullHandling = false, bool descending = false)
+        public static IFullComparer<T> OrderBy<T, TKey>(this ComparerBuilderFor<T> @this, Func<T?, TKey?> selector, Func<ComparerBuilderFor<TKey>, IComparer<TKey>> comparerFactory, bool specialNullHandling = false, bool descending = false)
         {
             _ = comparerFactory ?? throw new ArgumentNullException(nameof(comparerFactory));
             var comparer = comparerFactory(ComparerBuilder.For<TKey>());
@@ -64,7 +73,7 @@ namespace Nito.Comparers
         /// <param name="specialNullHandling">A value indicating whether <c>null</c> values are passed to <paramref name="selector"/>. If <c>false</c>, then <c>null</c> values are considered less than any non-<c>null</c> values and are not passed to <paramref name="selector"/>. This value is ignored if <typeparamref name="T"/> is a non-nullable type.</param>
         /// <param name="descending">A value indicating whether the sorting is done in descending order. If <c>false</c> (the default), then the sort is in ascending order.</param>
         /// <returns>A key comparer.</returns>
-        public static IFullComparer<T> OrderBy<T, TKey>(this ComparerBuilderFor<T> @this, Func<T, TKey> selector, IComparer<TKey>? keyComparer = null, bool specialNullHandling = false, bool descending = false)
+        public static IFullComparer<T> OrderBy<T, TKey>(this ComparerBuilderFor<T> @this, Func<T?, TKey?> selector, IComparer<TKey>? keyComparer = null, bool specialNullHandling = false, bool descending = false)
         {
             var selectComparer = new SelectComparer<T, TKey>(selector, keyComparer, specialNullHandling);
             return descending ? selectComparer.Reverse() : selectComparer;