+++ src/Nito.Comparers.Core/EqualityComparerBuilderFor.cs
@@ -3,6 +3,9 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using Nito.Comparers.Util;
 
+#pragma warning disable IDE0079
+#pragma warning disable IDE0060
+
 namespace Nito.Comparers
 {
     /// <summary>
@@ -36,6 +39,13 @@ namespace Nito.Comparers
         public static IFullEqualityComparer<T> Default<T>(this EqualityComparerBuilderFor<T> @this) => (IFullEqualityComparer<T>)EqualityComparerHelpers.NormalizeDefault<T>(null);
 
         /// <summary>
+        /// Gets a natural string equality comparer, which treats numeric sequences (0-9) as numeric.
+        /// </summary>
+        /// <param name="this">The equality comparer builder.</param>
+        /// <param name="comparison">The comparison type used to compare the text segments of the string (not used for numeric segments).</param>
+        public static IFullEqualityComparer<string> Natural(this EqualityComparerBuilderFor<string> @this, StringComparison comparison) => new NaturalStringComparer(comparison);
+
+        /// <summary>
         /// Gets the reference equality comparer for this type.
         /// </summary>
         public static IFullEqualityComparer<T> Reference<T>(this EqualityComparerBuilderFor<T> @this)
@@ -54,7 +64,7 @@ namespace Nito.Comparers
         /// <param name="comparerFactory">The definition of the key comparer. May not be <c>null</c>.</param>
         /// <param name="specialNullHandling">A value indicating whether <c>null</c> values are passed to <paramref name="selector"/>. If <c>false</c>, then <c>null</c> values are considered less than any non-<c>null</c> values and are not passed to <paramref name="selector"/>. This value is ignored if <typeparamref name="T"/> is a non-nullable type.</param>
         /// <returns>A key comparer.</returns>
-        public static IFullEqualityComparer<T> EquateBy<T, TKey>(this EqualityComparerBuilderFor<T> @this, Func<T, TKey> selector, Func<EqualityComparerBuilderFor<TKey>, IEqualityComparer<TKey>> comparerFactory, bool specialNullHandling = false)
+        public static IFullEqualityComparer<T> EquateBy<T, TKey>(this EqualityComparerBuilderFor<T> @this, Func<T?, TKey?> selector, Func<EqualityComparerBuilderFor<TKey>, IEqualityComparer<TKey>> comparerFactory, bool specialNullHandling = false)
         {
             _ = comparerFactory ?? throw new ArgumentNullException(nameof(comparerFactory));
             var comparer = comparerFactory(EqualityComparerBuilder.For<TKey>());
@@ -71,7 +81,7 @@ namespace Nito.Comparers
         /// <param name="keyComparer">The key comparer. Defaults to <c>null</c>. If this is <c>null</c>, the default comparer is used.</param>
         /// <param name="specialNullHandling">A value indicating whether <c>null</c> values are passed to <paramref name="selector"/>. If <c>false</c>, then <c>null</c> values are considered less than any non-<c>null</c> values and are not passed to <paramref name="selector"/>. This value is ignored if <typeparamref name="T"/> is a non-nullable type.</param>
         /// <returns>A key comparer.</returns>
-        public static IFullEqualityComparer<T> EquateBy<T, TKey>(this EqualityComparerBuilderFor<T> @this, Func<T, TKey> selector, IEqualityComparer<TKey>? keyComparer = null, bool specialNullHandling = false)
+        public static IFullEqualityComparer<T> EquateBy<T, TKey>(this EqualityComparerBuilderFor<T> @this, Func<T?, TKey?> selector, IEqualityComparer<TKey>? keyComparer = null, bool specialNullHandling = false)
         {
             return new SelectEqualityComparer<T, TKey>(selector, keyComparer, specialNullHandling);
         }