+++ src/Nito.Comparers.Core/Fixes/ExplicitGetHashCodeComparer.cs
@@ -15,16 +15,16 @@ namespace Nito.Comparers.Fixes
         /// </summary>
         /// <param name="source">The source comparer. If this is <c>null</c>, the default comparer is used.</param>
         /// <param name="getHashCode">The <c>GetHashCode</c> implementation to use. If this is <c>null</c>, this type will attempt to find <c>GetHashCode</c> on <paramref name="source"/>; if none is found, throws an exception.</param>
-        public ExplicitGetHashCodeComparer(IComparer<T>? source, Func<T, int>? getHashCode)
+        public ExplicitGetHashCodeComparer(IComparer<T>? source, Func<T?, int>? getHashCode)
             : base(source, getHashCode, false)
         {
         }
 
         /// <inheritdoc />
-        protected override int DoGetHashCode(T obj) => SourceGetHashCode(obj);
+        protected override int DoGetHashCode(T? obj) => SourceGetHashCode(obj);
 
         /// <inheritdoc />
-        protected override int DoCompare(T x, T y) => Source.Compare(x, y);
+        protected override int DoCompare(T? x, T? y) => Source.Compare(x!, y!);
 
         /// <summary>
         /// Returns a short, human-readable description of the comparer. This is intended for debugging and not for other purposes.