+++ src/Nito.Comparers.Core/Internals/ISubstringComparer.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Nito.Comparers.Internals
+{
+    public interface ISubstringComparer
+    {
+        int GetHashCode(string source, int offset, int length);
+        int Compare(string stringA, int offsetA, int lengthA, string stringB, int offsetB, int lengthB);
+    }
+}