+++ src/Nito.Comparers.Core/Util/DefaultComparer.cs
@@ -19,13 +19,13 @@ namespace Nito.Comparers.Util
         }
 
         /// <inheritdoc />
-        protected override int DoGetHashCode(T obj) => EqualityComparer<T>.Default.GetHashCode(obj!);
+        protected override int DoGetHashCode(T? obj) => EqualityComparer<T>.Default.GetHashCode(obj!);
 
         /// <inheritdoc />
-        protected override bool DoEquals(T x, T y) => EqualityComparer<T>.Default.Equals(x, y);
+        protected override bool DoEquals(T? x, T? y) => EqualityComparer<T>.Default.Equals(x!, y!);
 
         /// <inheritdoc />
-        protected override int DoCompare(T x, T y) => Comparer<T>.Default.Compare(x, y);
+        protected override int DoCompare(T? x, T? y) => Comparer<T>.Default.Compare(x!, y!);
 
         /// <summary>
         /// Gets the default comparer for this type.